Reset search box when searching with term
If you choose a category, all songs in that category will display, and the category will stay selected. This allows you to search within the category.

If you search with a search term in omni_search, upon execution of the search, the category will be reset back to 'All Categories' and the term in the search box will get removed, allowing easier access to search again.