Handle first-run installation
This checks for a hidden preference value to see if we've run the app before. If we have not, it attempts to do setup.

If it finds `~/mrvoice.cfg` or `C:\mrvoice.cfg`, it assumes that this system has an existing Mr. Voice 2 installation. In that case, it reads the `.cfg` file and sets the Mx. Voice preferences to point to the same database, music directory, and hotkey directory as Mr. Voice 2. That provides backward compatibility to run both.

If it does not, it uses the `[userData](https://www.electronjs.org/docs/api/app#appgetpathname)` directory on your system, and creates mp3 and hotkey directories in there. It also creates a `mxvoice.db` file, creates the schemas, and adds one category and song into the database. That way, you start with something.

If the hidden preference says that we've run Mx. Voice before, all of the above is skipped.

To test this: Remove your userData directory (on a Mac, it will be `~/Library/Application Support/Mx. Voice/`), rename your mrvoice.cfg temporarily, and then run the app. It should start and show you a database with one song and one category.

To revert back to a full database from Mr. Voice 2, remove the userData directory and make sure you rename your config file back to `~/mrvoice.cfg`. Then when you run it, it should recreate your preferences in the migrated manner and show you your previous Mr. Voice 2 data.