Convert from electron-preferences to electron-store
This removes the old `electron-preferences` module from the code, replacing it with the better-supported `electron-store` module.

Adds a check to see if we're using the old `preferences.json` file, or the new `config.json` file. If the old file is found, values are migrated and the old `preferences.json` file is renamed.

Needs some UI help on the modal