Remove remote module, fix dialog boxes
Something broke with Electron 9 in the electron-prompt boxes, so that it failed loading the custom stylesheets for coloring. This led to a refactor of using IPC to get the app path in the main process, then send it over to the renderer.

Also refactors away the remote module usage for the directory picker in preferences, since remote will go away at some point soon.