Log retention and export
This persists all logs generated to an internal file, and adds a new item into the main menu that allows a user to export the log data to a file. This can be useful if a user is having problems with their app, so that they can send us their logs for analysis.